fix(api): relocate release_layout into scripts/common (root-scripts gate)
The neutral layout module landed at scripts/ root, which the
'No new root scripts' gate correctly refuses. scripts/common/ is the
established neutral-shared package (common.thresholds precedent);
importers updated (path_safety, release_snapshot, tests).

X-Agent: claude-infra
