+++ scripts/common/release_layout.py
